create useFilterData to offload state into hook for cryptotable sorting
